tools: refactor dep_updaters
PR-URL: https://github.com/nodejs/node/pull/46488

+++ tools/dep_updaters/update-eslint.sh
@@ -7,17 +7,26 @@
 
 set -ex
 
+ROOT=$(cd "$(dirname "$0")/../.." && pwd)
+
+[ -z "$NODE" ] && NODE="$ROOT/out/Release/node"
+[ -x "$NODE" ] || NODE=$(command -v node)
+NPM="$ROOT/deps/npm/bin/npm-cli.js"
+
+NEW_VERSION=$("$NODE" "$NPM" view eslint dist-tags.latest)
+CURRENT_VERSION=$("$NODE" -p "require('./tools/node_modules/eslint/package.json').version")
+
+if [ "$NEW_VERSION" = "$CURRENT_VERSION" ]; then
+  echo "Skipped because ESlint is on the latest version."
+  exit 0
+fi
+
 cd "$( dirname "$0" )" || exit
 rm -rf ../node_modules/eslint
 (
     rm -rf eslint-tmp
     mkdir eslint-tmp
     cd eslint-tmp || exit
-
-    ROOT="$PWD/../../.."
-    [ -z "$NODE" ] && NODE="$ROOT/out/Release/node"
-    [ -x "$NODE" ] || NODE=$(command -v node)
-    NPM="$ROOT/deps/npm/bin/npm-cli.js"
 
     "$NODE" "$NPM" init --yes
 
@@ -63,3 +72,7 @@ rm -rf ../node_modules/eslint
 
 mv eslint-tmp/node_modules/eslint ../node_modules/eslint
 rm -rf eslint-tmp/
+
+# The last line of the script should always print the new version,
+# as we need to add it to $GITHUB_ENV variable.
+echo "NEW_VERSION=$NEW_VERSION"
